v2.14.4
refactor(oracle): extract OCI classes to axon::database2r::oci namespace

- oci.h/oci.cpp: new files containing environment, error, context,
  server,
  session, connection, statement, column — all extracted from
oracle.h/oracle.cpp
  and moved to axon::database2r::oci namespace

- oracle.h: remove environment, error, context, server, session,
  statement,
  column nested classes; replace with oci:: equivalents; remove
redundant
  _environment, _connected, _executed members; _connection now
  shared_ptr<oci::connection>; add oracle(shared_ptr<oci::connection>)
ctor
  for shared connection use case

- oracle.cpp: remove all static definitions for environment (now in
  oci.cpp);
  remove statement implementation (now in oci.cpp); update all _context,
  _server, _session, _error usages to _connection->ctx() /
_connection->error();
  fix missing throw in connect(); fix _get_double() dangling pointer;
  fix _get_string() find_last_not_of UB on all-space strings; add
  OCI_SUCCESS_WITH_INFO truncation warning in fetch(); add rlen.resize
  for SQLT_BFLOAT/SQLT_BDOUBLE

- util.h/util.cpp: add validator::tns() and regex_tns for TNS name
  validation; change hostname() and username() to take const ref

- CMakeLists: add oci.cpp to build

- examples/oracle.cpp: add environment guard; add shared connection test
  via oracle(shared_ptr<oci::connection>)

version: 2.14.3-8710 -> 2.14.4-8822
