+++ src/CMakeLists.txt
@@ -18,6 +18,7 @@ list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/md5.cpp)
 list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/message.cpp)
 list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/mml.cpp)
 list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/nothing.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/oci.cpp)
 list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/oracle.cpp)
 list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/rabbit.cpp)
 list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/recordset.cpp)
