+++ src/util.cpp
@@ -33,6 +33,7 @@ namespace axon
 
 		const boost::regex validator::regex_ipaddr{"(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)(.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)){3}"};
 		const boost::regex validator::regex_fqdn{"^(([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9-]*[a-zA-Z0-9]).)*([A-Za-z0-9]|[A-Za-z0-9][A-Za-z0-9-]*[A-Za-z0-9])$"};
+		const boost::regex validator::regex_tns{"^[a-zA-Z][a-zA-Z0-9_\\.-]{0,29}$"};
 		const boost::regex validator::regex_username{"^[A-Za-z0-9]+(?:[._-][A-Za-z0-9]+)*$"};
 
 		unsigned long long bytes_to_ull(const char* bytes, size_t size)
